Remove ImageContent and AudioContent
